refactor: move RacingApplication to controller package, change View parameter with CarCollection
