+++ src/main/java/racing/RacingMain.java
@@ -1,5 +1,7 @@
 package racing;
 
+import racing.controller.RacingApplication;
+
 public class RacingMain {
     public static void main(String[] args) {
         new RacingApplication().run();
